Add PV1/PV2 power and voltage support from Home Assistant
- Add sensor mappings for pv1_power, pv2_power, pv1_voltage, pv2_voltage
- Update DeviceData model to use double for voltage fields
- Now displays separate PV1 and PV2 data on website

+++ LumenTreeInfo.Lib/SolarInverterMonitor.cs
@@ -852,9 +852,9 @@ public class SolarInverterMonitor : IDisposable
         public int? HomeLoad { get; set; }
 
         // PV (Solar) Input
-        public int? Pv1Voltage { get; set; }
+        public double? Pv1Voltage { get; set; }
         public int? Pv1Power { get; set; }
-        public int? Pv2Voltage { get; set; }
+        public double? Pv2Voltage { get; set; }
         public int? Pv2Power { get; set; }
         public int? TotalPvPower { get; set; }
